Added file by upload
Solves CodeFights sample tests. Does not solve full tests.
I am unable to see the full test data in order to diagnose the problem.
It is also annoying that I can only see the function I submit for the solution. I cannot see the program in which it is embedded.